requirements.txt
Images should be in .jpg, .jpeg or .png format
The Project needs Python 3.8 or higher
Following Libraries are used:
    numpy==1.24.3
    Pillow==9.5.0
    Pillow==9.4.0
    Pillow==10.1.0
    Requests==2.31.0
    scikit_learn==1.2.2
    scikit_learn==1.3.0
    sentence_transformers==2.2.2

How to run:
1. Update the url_list in main.py.
2. Run -> python main.py 
3. Output will get printed.


